Add files via upload
Dynamic Host Configuration Protocol (DHCP) is a client/server protocol that automatically provides an Internet Protocol (IP) host with its IP address and other related configuration information such as the subnet mask and default gateway. DHCP is an application layer protocol.In this project we have used Cisco Packet Tracer tool for dynamically allocating IP addresses to the devices connected in the network.